refactor: reorganize packages for user interactions (no behavior change)
