+++ itachallenge-user/src/main/java/com/itachallenge/App.java
@@ -1,0 +1,20 @@
+package com.itachallenge;
+
+import com.itachallenge.githubcore.config.GithubServiceConfig;
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.Import;
+
+@SpringBootApplication
+@Import(GithubServiceConfig.class)
+@EnableDiscoveryClient
+@OpenAPIDefinition(info = @Info(title = "Ita Backend User", version = "1.0", description = "Description"))
+public class App {
+
+    public static void main(String[] args) {
+        SpringApplication.run(App.class, args);
+    }
+}
