+++ itachallenge-user/src/main/java/com/itachallenge/user/controller/userinteraction/bookmark/BookmarkController.java
@@ -1,0 +1,5 @@
+package com.itachallenge.user.controller.userinteraction.bookmark;
+
+public class BookmarkController {
+    //to do
+}
