+++ itachallenge-user/src/main/java/com/itachallenge/user/controller/userinteraction/favorite/FavoriteController.java
@@ -1,0 +1,5 @@
+package com.itachallenge.user.controller.userinteraction.favorite;
+
+public class FavoriteController {
+    //to do
+}
