+++ itachallenge-user/src/main/java/com/itachallenge/user/dto/userinteraction/bookmark/BookmarkDto.java
@@ -1,0 +1,5 @@
+package com.itachallenge.user.dto.userinteraction.bookmark;
+
+public class BookmarkDto {
+
+}
