+++ itachallenge-user/src/main/java/com/itachallenge/user/dto/userinteraction/favorite/FavoriteDto.java
@@ -1,0 +1,5 @@
+package com.itachallenge.user.dto.userinteraction.favorite;
+
+public class FavoriteDto {
+
+}
