+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/document/bookmark/Bookmark.java
@@ -1,0 +1,5 @@
+package com.itachallenge.userinteraction.document.bookmark;
+
+public class Bookmark {
+
+}
