+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/document/favorite/Favorite.java
@@ -1,0 +1,5 @@
+package com.itachallenge.userinteraction.document.favorite;
+
+public class Favorite {
+
+}
