+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/repository/bookmark/BookmarkRepository.java
@@ -1,0 +1,5 @@
+package com.itachallenge.userinteraction.repository.bookmark;
+
+public class BookmarkRepository {
+    //to do
+}
