+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/repository/favorite/FavoriteRepository.java
@@ -1,0 +1,4 @@
+package com.itachallenge.userinteraction.repository.favorite;
+
+public class FavoriteRepository {
+}
