+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/service/bookmark/BookmarkService.java
@@ -1,0 +1,5 @@
+package com.itachallenge.userinteraction.service.bookmark;
+
+public interface BookmarkService {
+    //to do
+}
