+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/service/favorite/FavoriteService.java
@@ -1,0 +1,5 @@
+package com.itachallenge.userinteraction.service.favorite;
+
+public interface FavoriteService {
+    //to do
+}
