--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/controller/bookmark/BookmarkController.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.controller.bookmark;
-
-public class BookmarkController {
-    //to do
-}
