--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/controller/favourite/FavouriteController.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.controller.favourite;
-
-public class FavouriteController {
-    //to do
-}
