--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/document/bookmark/Bookmark.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.document.bookmark;
-
-public class Bookmark {
-
-}
