--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/document/favourite/Favourite.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.document.favourite;
-
-public class Favourite {
-
-}
