--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/dto/bookmark/BookmarkDto.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.dto.bookmark;
-
-public class BookmarkDto {
-
-}
