--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/dto/common/ErrorResponse.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.dto.common;
-
-public class ErrorResponse {
-
-}
