--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/dto/favourite/FavouriteDto.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.dto.favourite;
-
-public class FavouriteDto {
-
-}
