--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/event/contract/UserFavouriteChanged.java
@@ -1,4 +1,0 @@
-package com.itachallenge.user.interactions.events.contract;
-
-public class UserFavouriteChangedV1 {
-}
