--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/event/publisher/InteractionEventPublisher.java
@@ -1,4 +1,0 @@
-package com.itachallenge.user.interactions.events.publisher;
-
-public class InteractionEventPublisher {
-}
