--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/repository/bookmark/BookmarkRepository.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.repository.bookmark;
-
-public class BookmarkRepository {
-    //to do
-}
