--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/repository/favourite/FavouriteRepository.java
@@ -1,4 +1,0 @@
-package com.itachallenge.user.interactions.repository.favourite;
-
-public class FavouriteRepository {
-}
