--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/service/bookmark/BookmarkService.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.service.bookmark;
-
-public interface BookmarkService {
-    //to do
-}
