--- itachallenge-user/user-interactions/src/main/java/com/itchallenge/userInteractions/service/favourite/FavouriteService.java
@@ -1,5 +1,0 @@
-package com.itachallenge.user.interactions.service.favourite;
-
-public interface FavouriteService {
-    //to do
-}
